refactor: club dto 패키징 수정
